Use new ArrowHead class for MultiArrow
